pyproject.toml
[tool.poetry]
name = "django-challenge-001"
version = "0.1.0"
description = ""
authors = ["Lucas Pinheiro <lucas4pinheiro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.1.5"
djangorestframework = "3.12.2"
psycopg2 = "2.8.6"
django-environ = "^0.4.5"
Pillow = "^8.2.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"
flake8 = "^3.9.2"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"